feat: reintroduce subclassing
