+++ src/structures/CommandInteraction.js
@@ -1,50 +1,15 @@
 'use strict';
 
+const BaseCommandInteraction = require('./BaseCommandInteraction');
 const CommandInteractionOptionResolver = require('./CommandInteractionOptionResolver');
-const Interaction = require('./Interaction');
-const InteractionWebhook = require('./InteractionWebhook');
-const InteractionResponses = require('./interfaces/InteractionResponses');
-const { ApplicationCommandOptionTypes } = require('../util/Constants');
 
 /**
  * Represents a command interaction.
- * @extends {Interaction}
- * @implements {InteractionResponses}
+ * @extends {BaseCommandInteraction}
  */
-class CommandInteraction extends Interaction {
+class CommandInteraction extends BaseCommandInteraction {
   constructor(client, data) {
     super(client, data);
-
-    /**
-     * The channel this interaction was sent in
-     * @type {?TextBasedChannels}
-     * @name CommandInteraction#channel
-     * @readonly
-     */
-
-    /**
-     * The id of the channel this interaction was sent in
-     * @type {Snowflake}
-     * @name CommandInteraction#channelId
-     */
-
-    /**
-     * The invoked application command's id
-     * @type {Snowflake}
-     */
-    this.commandId = data.data.id;
-
-    /**
-     * The invoked application command's name
-     * @type {string}
-     */
-    this.commandName = data.data.name;
-
-    /**
-     * Whether the reply to this interaction has been deferred
-     * @type {boolean}
-     */
-    this.deferred = false;
 
     /**
      * The options passed to the command.
@@ -52,136 +17,9 @@ class CommandInteraction extends Interaction {
      */
     this.options = new CommandInteractionOptionResolver(
       this.client,
-      typeof data.data.target_id !== 'undefined'
-        ? this.resolveContextMenuOptions(data.data)
-        : data.data.options?.map(option => this.transformOption(option, data.data.resolved)) ?? [],
+      data.data.options?.map(option => this.transformOption(option, data.data.resolved)) ?? [],
     );
-
-    /**
-     * Whether this interaction has already been replied to
-     * @type {boolean}
-     */
-    this.replied = false;
-
-    /**
-     * Whether the reply to this interaction is ephemeral
-     * @type {?boolean}
-     */
-    this.ephemeral = null;
-
-    /**
-     * An associated interaction webhook, can be used to further interact with this interaction
-     * @type {InteractionWebhook}
-     */
-    this.webhook = new InteractionWebhook(this.client, this.applicationId, this.token);
   }
-
-  /**
-   * The invoked application command, if it was fetched before
-   * @type {?ApplicationCommand}
-   */
-  get command() {
-    const id = this.commandId;
-    return this.guild?.commands.cache.get(id) ?? this.client.application.commands.cache.get(id) ?? null;
-  }
-
-  /**
-   * Represents an option of a received command interaction.
-   * @typedef {Object} CommandInteractionOption
-   * @property {string} name The name of the option
-   * @property {ApplicationCommandOptionType} type The type of the option
-   * @property {string|number|boolean} [value] The value of the option
-   * @property {CommandInteractionOption[]} [options] Additional options if this option is a
-   * subcommand (group)
-   * @property {User} [user] The resolved user
-   * @property {GuildMember|APIInteractionDataResolvedOption} [member] The resolved member
-   * @property {GuildChannel|APIInteractionDataResolvedOption} [channel] The resolved channel
-   * @property {Role|APIRole} [role] The resolved role
-   */
-
-  /**
-   * Transforms an option received from the API.
-   * @param {APIApplicationCommandOption} option The received option
-   * @param {APIApplicationCommandOptionResolved} resolved The resolved interaction data
-   * @returns {CommandInteractionOption}
-   * @private
-   */
-  transformOption(option, resolved) {
-    const result = {
-      name: option.name,
-      type: ApplicationCommandOptionTypes[option.type],
-    };
-
-    if ('value' in option) result.value = option.value;
-    if ('options' in option) result.options = option.options.map(opt => this.transformOption(opt, resolved));
-
-    if (resolved) {
-      const user = resolved.users?.[option.value];
-      if (user) result.user = this.client.users._add(user);
-
-      const member = resolved.members?.[option.value];
-      if (member) result.member = this.guild?.members._add({ user, ...member }) ?? member;
-
-      const channel = resolved.channels?.[option.value];
-      if (channel) {
-        result.channel = this.client.channels._add(channel, this.guild, { fromInteraction: true }) ?? channel;
-      }
-
-      const role = resolved.roles?.[option.value];
-      if (role) result.role = this.guild?.roles._add(role) ?? role;
-    }
-
-    return result;
-  }
-
-  /**
-   * Resolves and transforms options received from the API for a context menu interaction.
-   * @param {APIApplicationCommandInteractionData} data The interaction data
-   * @returns {CommandInteractionOption[]}
-   * @private
-   */
-  resolveContextMenuOptions({ target_id, resolved }) {
-    const result = [];
-
-    if (resolved.users?.[target_id]) {
-      result.push(
-        this.transformOption({ name: 'user', type: ApplicationCommandOptionTypes.USER, value: target_id }, resolved),
-      );
-    }
-
-    if (resolved.messages?.[target_id]) {
-      result.push({
-        name: 'message',
-        type: '_MESSAGE',
-        value: target_id,
-        message: this.channel?.messages._add(resolved.messages[target_id]),
-      });
-    }
-
-    return result;
-  }
-
-  // These are here only for documentation purposes - they are implemented by InteractionResponses
-  /* eslint-disable no-empty-function */
-  deferReply() {}
-  reply() {}
-  fetchReply() {}
-  editReply() {}
-  deleteReply() {}
-  followUp() {}
 }
 
-InteractionResponses.applyToClass(CommandInteraction, ['deferUpdate', 'update']);
-
 module.exports = CommandInteraction;
-
-/* eslint-disable max-len */
-/**
- * @external APIApplicationCommandOptionResolved
- * @see {@link https://discord.com/developers/docs/interactions/slash-commands#interaction-applicationcommandinteractiondataresolved}
- */
-
-/**
- * @external APIInteractionDataResolvedOption
- * @see {@link https://discord.com/developers/docs/interactions/slash-commands#sample-application-command-interaction-application-command-interaction-data-resolved-structure}
- */
